🔨 refactor: move mocks to a folder for consistency
- rename some files
