+++ server/src/domain/services/csv-service.ts
@@ -1,5 +1,5 @@
 import type { Transform } from 'node:stream'
 
-export interface CSVExportService {
+export interface CSVService {
   createCSV(columns: Array<{ key: string; header: string }>): Transform
 }
