+++ server/src/infra/services/node-csv-export-service.ts
@@ -1,8 +1,8 @@
 import type { Transform } from 'node:stream'
 import { stringify } from 'csv-stringify'
-import type { CSVExportService } from '@/domain/services/csv-export-service'
+import type { CSVService } from '@/domain/services/csv-service'
 
-export class NodeCSVExportService implements CSVExportService {
+export class NodeCSVService implements CSVService {
   createCSV(columns: Array<{ key: string; header: string }>): Transform {
     return stringify({
       delimiter: ',',
